Restructured codebase.
The codebase has been split in 4 different modules.
The code itself should be mostly the same, as I didn't wan't to introduce new potential bugs to the code.

Instructions on how to run the code can be found in the README.

To add new tests, add them to the module folder and call the file `test_module_name.py`. Note the *test_*, this is important, otherwise pytest won't find the file.

For information on how the application is structured, check https://www.digitalocean.com/community/tutorials/how-to-structure-large-flask-applications#structuring-the-application-directory and take a look at the existing files. If you still have questions, send me a message.

On the different request methods:

`GET` requests are for showing data
`POST` requests are for creating data
`PUT` requests are for updating data
`DELETE` requests are for deleting data

HTML forms only support `GET` and `POST` requests (i.e. `<form action="/route" method="get|post">` are the only working forms) so instead use an AJAX request. For an example on how that works, take a look at dataset-view.html or table-view.html and/or send me a message.

Something I noticed wasn't working yet:

* deleting schemas - When clicking the delete schema button, after reloading the page the schema is still there but it's 'empty' (name and description are `None`)

On implementing an API:

We could implement API routes that have the same functionality as the routes we currently have (e.g. `/api/login` does what `/login` already does) and then point our current routes to those API routes and 'translate' the user data (e.g. the `/login` route takes the info the user submitted form and makes an API call with that same info in a query). I'd love to have your input on this subject, so please either send me a message about it or bring it up during the next meeting.

Some handy links about designing APIs:

* [A guide on creating a RESTful API using Flask *without* Flask-Restful](https://blog.miguelgrinberg.com/post/designing-a-restful-api-with-python-and-flask)
* [A guide on creating a RESTful API using Flask *with* Flask-Restful](http://blog.luisrei.com/articles/flaskrest.html)
