+++ README.md
@@ -9,6 +9,24 @@ The goal of this project is to create a webapp that provides easy to use data ma
 Users can upload their data (from various file formats), share access and divide it into logical 'datasets'.
 Furthermore, the app allows for data manipulation in order to turn raw data into usable information. Eventually, the idea is to provide sufficiently advanced data cleaning and analysis tools.
 
+### Usage
+
+*Requirements*
+
+* pip
+* virtualenv
+* PostgreSQL
+
+To setup your system, run:
+`./setup.sh`
+
+To run the tests, run:
+`pytest`
+
+To run the app, run:
+`python3 run.py`
+
+
 ### Team
 This project is being realised by four undergraduate students at the University of Antwerp:
