--- app/default_config.py
@@ -1,6 +1,0 @@
-config_data = dict()
-config_data['app_name'] = 'Project_Databases'
-config_data['dbname'] = 'userdb'
-config_data['dbuser'] = 'dbadmin'
-config_data['dbpass'] = 'dbadmin'
-config_data['dbhost'] = 'localhost'
