+++ app/templates/data_service/dataset-view.html
@@ -27,11 +27,22 @@
                     <tbody>
                     {% for t in tables %}
                         <tr>
-                            <td><a href="/data-service/{{ ds.id }}/{{ t.name }}">{{ t.name }}</a></td>
+                            <td><a href="/datasets/{{ ds.id }}/{{ t.name }}">{{ t.name }}</a></td>
                             <td> {{ t.desc }}</td>
-                            <td style="text-align: center;"><a href="/data-service/{{ ds.id }}/{{ t.name }}">
-                                <button type="button" class="btn btn-xs btn-danger">❌</button>
-                            </a></td>
+                            <td style="text-align: center;">
+                                <form action="/datasets/{{ ds.id }}/{{ t.name }}">
+                                    <button type="button" class="btn btn-xs btn-danger" onclick="deleteRequest()">❌
+                                    </button>
+                                </form>
+                                <script>
+                                    function deleteRequest() {
+                                        $.ajax({
+                                            type: 'DELETE',
+                                            url: $("form").attr("action")
+                                        });
+                                    }
+                                </script>
+                            </td>
                         </tr>
                     {% endfor %}
                     </tbody>
@@ -46,8 +57,8 @@
         <div class="col-sm-2"></div>
         <div class="col-sm-8">
             <h4>Add CSV/Zip</h4>
-            <form action="/data-service/{{ ds.id }}" method="post" enctype=multipart/form-data>
-                <input type="file" name="file" id="ds-files" accept=".csv,.zip" style='margin-bottom: 10px;'>
+            <form action="/datasets/{{ ds.id }}" method="post" enctype=multipart/form-data>
+                <input type="file" name="file" id="ds-files" accept=".csv,.zip,.dump" style='margin-bottom: 10px;'>
                 <br>
                 <button type="submit" class="btn btn-default" id="submit">Submit</button>
             </form>
