+++ requirements.txt
@@ -1,3 +1,4 @@
+attrs==17.4.0
 click==6.7
 Flask==0.12.2
 Flask-Login==0.4.1
@@ -5,6 +6,10 @@ itsdangerous==0.24
 Jinja2==2.10
 MarkupSafe==1.0
 passlib==1.7.1
+pluggy==0.6.0
 psycopg2==2.7.4
 psycopg2-binary==2.7.4
+py==1.5.2
+pytest==3.4.2
+six==1.11.0
 Werkzeug==0.14.1
